pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sediment-palace-mcp"
version = "0.1.0"
description = "Local-first memory MCP server with sedimentation model"
readme = "project/PROJECT_OVERVIEW.md"
requires-python = ">=3.11"
dependencies = [
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q -p no:cacheprovider"
